+++ reto_Guardians/Controllers/RelacionesPController.cs
@@ -18,8 +18,11 @@ namespace reto_Guardians.Controllers
             _context = context;
         }
 
-        // GET: api/RelacionesPersonales
+        // GET: api/RelacionesP
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(RelacionPDTO))]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<IEnumerable<RelacionPDTO>>> GetRelacionesP()
         {
             try
@@ -49,10 +52,18 @@ namespace reto_Guardians.Controllers
 
         // GET: api/RelacionesPersonales/BuscarRelacionId/1
         [HttpGet("BuscarRelacionId/{idrp}")]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(RelacionPDTO))]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult<RelacionPDTO>> GetRelacionP(int idrp)
         {
             try
             {
+                if (idrp<=0)
+                {
+                    return BadRequest("El id de la relación no puede ser 0 o menor a este");
+                }
                 var rp = await _context.RelacionesPersonales
                 .FirstOrDefaultAsync(p => p.IdPersona1 == idrp || p.IdPersona2==idrp);
                 if (rp == null)
@@ -77,8 +88,16 @@ namespace reto_Guardians.Controllers
 
         // PUT: api/RelacionesPersonales/ModificarRelacion/5
         [HttpPut("ModificarRelacion/{idrp}")]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(RelacionPDTO))]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> PutRelacionP(int idrp, [FromBody] RelacionPDTO rpUpdate)
         {
+            if (idrp <= 0)
+            {
+                return BadRequest("El id de la relación no puede ser 0 o menor a este");
+            }
             var existingRp = await _context.RelacionesPersonales.FindAsync(idrp);
             if (existingRp == null)
             {
@@ -115,15 +134,24 @@ namespace reto_Guardians.Controllers
         }
         // POST: api/RelacionesPersonales/CrearRelacion
         [HttpPost("CrearRelacion")]
-        public async Task<ActionResult<RelacionPersonal>> PostPersona([FromBody] RelacionPDTO nueva)
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(PersonaDTO))]
+        [ProducesResponseType(StatusCodes.Status409Conflict)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<RelacionPersonal>> PostRelacionP([FromBody] RelacionPDTO nueva)
         {
             try
             {
+                var relacionExistente = await _context.RelacionesPersonales
+                .FirstOrDefaultAsync(a => a.TipoRelacion == nueva.TipoRelacion && a.IdPersona1 == nueva.Id1 && a.IdPersona2==nueva.Id2);
+                if (relacionExistente != null)
+                {
+                    return Conflict($"La relación {nueva.TipoRelacion} entre {nueva.Id1} y {nueva.Id2} ya existe");
+                }
                 var relacionNueva = new RelacionPersonal
                 {
                     TipoRelacion = nueva.TipoRelacion ?? string.Empty,
-                    IdPersona1 = nueva.Id1.HasValue ? nueva.Id1.Value : 0,
-                    IdPersona2 = nueva.Id2.HasValue ? nueva.Id2.Value : 0
+                    IdPersona1 = nueva.Id1 ?? 0,
+                    IdPersona2 = nueva.Id2 ?? 0
                 };
                 _context.RelacionesPersonales.Add(relacionNueva);
                 await _context.SaveChangesAsync();
@@ -137,16 +165,24 @@ namespace reto_Guardians.Controllers
 
         // DELETE: api/RelacionesPersonales/BorrarRelacion/5
         [HttpDelete("BorrarRelacion/{idpersona}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> DeleteRelacion(int idrp)
         {
             if (_context.RelacionesPersonales == null)
             {
-                return NotFound();
+                return NotFound("No hay registros");
+            }
+            if (idrp <= 0)
+            {
+                return BadRequest("El id de la relación no puede ser 0 o menor a este");
             }
             var relacion = await _context.RelacionesPersonales.Where(a => a.IdRelacion == idrp).FirstAsync();
             if (relacion == null)
             {
-                return NotFound();
+                return NotFound($"No se encontró la relación con id {idrp}");
             }
             _context.RelacionesPersonales.Remove(relacion);
             await _context.SaveChangesAsync();
